+++ assets/javascripts/kitten/components/tables/list-table/test.js
@@ -5,12 +5,15 @@ import { ListTable } from './index'
 describe('<ListTable />', () => {
   let component
 
-  describe('with a complete structure', () => {
+  describe('with a complete structure and custom classes', () => {
     beforeEach(() => {
       component = renderer
         .create(
           <ListTable id="TestListTable">
-            <ListTable.Header>
+            <ListTable.Header
+              className="customHeaderClass"
+              listProps={{ className: 'customListClass' }}
+            >
               <ListTable.Col>Header 1</ListTable.Col>
               <ListTable.Col>Header 2</ListTable.Col>
               <ListTable.Col>Header 3</ListTable.Col>
@@ -22,7 +25,10 @@ describe('<ListTable />', () => {
                 <ListTable.Col>Content 2</ListTable.Col>
                 <ListTable.Col>Content 3</ListTable.Col>
               </ListTable.Row>
-              <ListTable.Row>
+              <ListTable.Row
+                className="customRowClass"
+                listProps={{ className: 'customListClass' }}
+              >
                 <ListTable.Col>Content 1</ListTable.Col>
                 <ListTable.Col>Content 2</ListTable.Col>
                 <ListTable.Col>Content 3</ListTable.Col>